ISPN-6131 Properly handle script cache remove/clear ops
