+++ scripting/src/test/java/org/infinispan/scripting/ScriptingTest.java
@@ -1,16 +1,21 @@
 package org.infinispan.scripting;
 
 import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertNotNull;
+import static org.testng.AssertJUnit.assertTrue;
 
 import java.io.IOException;
 import java.io.InputStream;
 
+import org.infinispan.commons.CacheException;
 import org.infinispan.manager.EmbeddedCacheManager;
 import org.infinispan.tasks.TaskContext;
 import org.infinispan.test.SingleCacheManagerTest;
 import org.infinispan.test.TestingUtil;
 import org.infinispan.test.fwk.TestCacheManagerFactory;
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
+import static org.infinispan.scripting.ScriptingTests.loadScript;
 
 @Test(groups = "functional", testName = "scripting.ScriptingTest")
 public class ScriptingTest extends SingleCacheManagerTest {
@@ -22,28 +27,25 @@ public class ScriptingTest extends SingleCacheManagerTest {
       return TestCacheManagerFactory.createCacheManager();
    }
 
-   protected String[] getScripts() {
-      return new String[] { "test.js" };
-   }
-
    @Override
    protected void setup() throws Exception {
       super.setup();
       scriptingManager = cacheManager.getGlobalComponentRegistry().getComponent(ScriptingManager.class);
-      for (String scriptName : getScripts()) {
-         loadScript(scriptingManager, scriptName);
-      }
    }
 
-   public static void loadScript(ScriptingManager scriptingManager, String scriptName) throws IOException {
-      try (InputStream is = ScriptingTest.class.getResourceAsStream("/" + scriptName)) {
-         String script = TestingUtil.loadFileAsString(is);
-         scriptingManager.addScript(scriptName, script);
-      }
-   }
+
 
    public void testSimpleScript() throws Exception {
+      loadScript(scriptingManager, "test.js");
       String result = (String) scriptingManager.runScript("test.js", new TaskContext().addParameter("a", "a")).get();
       assertEquals("a", result);
+   }
+
+   @Test(expectedExceptions = CacheException.class, expectedExceptionsMessageRegExp = "ISPN026005:.*")
+   public void testScriptRemove() throws Exception {
+      loadScript(scriptingManager, "test.js");
+      scriptingManager.getScript("test.js");
+      scriptingManager.removeScript("test.js");
+      scriptingManager.getScript("test.js");
    }
 }
