+++ scripting/src/test/java/org/infinispan/scripting/SecureScriptingTest.java
@@ -1,7 +1,10 @@
 package org.infinispan.scripting;
 
+import static org.infinispan.scripting.ScriptingTests.loadScript;
 import static org.testng.AssertJUnit.assertEquals;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.security.PrivilegedAction;
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
@@ -19,16 +22,18 @@ import org.infinispan.security.AuthorizationPermission;
 import org.infinispan.security.Security;
 import org.infinispan.security.impl.IdentityRoleMapper;
 import org.infinispan.tasks.TaskContext;
+import org.infinispan.test.SingleCacheManagerTest;
 import org.infinispan.test.TestingUtil;
 import org.infinispan.test.fwk.TestCacheManagerFactory;
 import org.testng.annotations.Test;
 
 @Test(groups = "functional", testName = "scripting.SecureScriptingTest")
-public class SecureScriptingTest extends ScriptingTest {
+public class SecureScriptingTest extends SingleCacheManagerTest {
 
    static final Subject ADMIN = TestingUtil.makeSubject("admin", ScriptingManagerImpl.SCRIPT_MANAGER_ROLE);
    static final Subject RUNNER = TestingUtil.makeSubject("runner", "runner");
    static final Subject PHEIDIPPIDES = TestingUtil.makeSubject("pheidippides", "pheidippides");
+   private ScriptingManager scriptingManager;
 
    @Override
    protected EmbeddedCacheManager createCacheManager() throws Exception {
@@ -52,19 +57,13 @@ public class SecureScriptingTest extends ScriptingTest {
       return TestCacheManagerFactory.createCacheManager(global, config);
    }
 
-
-
-   @Override
-   protected String[] getScripts() {
-      return new String[] { "test.js", "testRole.js" };
-   }
-
    @Override
    protected void setup() throws Exception {
       Security.doAs(ADMIN, new PrivilegedExceptionAction<Void>() {
          @Override
          public Void run() throws Exception {
             SecureScriptingTest.super.setup();
+            scriptingManager = cacheManager.getGlobalComponentRegistry().getComponent(ScriptingManager.class);
             return null;
          }
       });
@@ -86,20 +85,31 @@ public class SecureScriptingTest extends ScriptingTest {
       Security.doAs(ADMIN, new PrivilegedAction<Void>() {
          @Override
          public Void run() {
-            cacheManager.getCache().clear();
+            SecureScriptingTest.super.clearContent();
             return null;
          }
       });
    }
 
-   @Override
+   public static void loadScript(ScriptingManager scriptingManager, String scriptName) throws Exception {
+      Security.doAs(ADMIN, new PrivilegedExceptionAction<Object>() {
+         @Override
+         public Void run() throws Exception {
+            loadScript(scriptingManager, scriptName);
+            return null;
+         }
+      });
+   }
+
    @Test(expectedExceptions= { SecurityException.class, CacheException.class} )
    public void testSimpleScript() throws Exception {
+      loadScript(scriptingManager, "test.js");
       String result = (String) scriptingManager.runScript("test.js", new TaskContext().addParameter("a", "a").cache(cache())).get();
       assertEquals("a", result);
    }
 
    public void testSimpleScriptWithEXECPermissions() throws Exception {
+      loadScript(scriptingManager, "test.js");
       String result = Security.doAs(RUNNER, new PrivilegedExceptionAction<String>() {
          @Override
          public String run() throws Exception {
@@ -111,6 +121,7 @@ public class SecureScriptingTest extends ScriptingTest {
 
    @Test(expectedExceptions= { PrivilegedActionException.class, CacheException.class} )
    public void testSimpleScriptWithEXECPermissionsWrongRole() throws Exception {
+      loadScript(scriptingManager, "testRole.js");
       String result = Security.doAs(RUNNER, new PrivilegedExceptionAction<String>() {
          @Override
          public String run() throws Exception {
@@ -121,6 +132,7 @@ public class SecureScriptingTest extends ScriptingTest {
    }
 
    public void testSimpleScriptWithEXECPermissionsRightRole() throws Exception {
+      loadScript(scriptingManager, "testRole.js");
       String result = Security.doAs(PHEIDIPPIDES, new PrivilegedExceptionAction<String>() {
          @Override
          public String run() throws Exception {
